requirements.txt
openai-whisper==20231117
fastapi==0.111.0
aiofiles==23.2.1
python-multipart
numpy
